+++ src/impls/is-string/__tests__/is-string.spec.ts
@@ -1,0 +1,9 @@
+import assert from 'static-type-assert';
+
+import { isString } from 'iter-tools-es';
+
+declare const value: any;
+
+if (isString(value)) {
+  assert<string>(value);
+}