+++ src/impls/not-string/__tests__/not-string.spec.ts
@@ -1,0 +1,9 @@
+import assert from 'static-type-assert';
+
+import { notString } from 'iter-tools-es';
+
+declare const value: string | number;
+
+if (notString(value)) {
+  assert<number>(value);
+}