+++ src/impls/not-string/not-string.js
@@ -1,0 +1,3 @@
+export function notString(value) {
+  return typeof value !== 'string';
+}